+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/RatifySpec.hs
@@ -37,7 +37,10 @@ import Test.Cardano.Ledger.Imp.Common
 
 spec ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 spec = do
   votingSpec
@@ -69,7 +72,10 @@ spec = do
 
 initiateHardForkWithLessThanMinimalCommitteeSize ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 initiateHardForkWithLessThanMinimalCommitteeSize =
   it "Hard Fork can still be initiated with less than minimal committee size" $ do
@@ -97,7 +103,10 @@ initiateHardForkWithLessThanMinimalCommitteeSize =
 
 spoAndCCVotingSpec ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 spoAndCCVotingSpec = do
   describe "When CC expired" $ do
@@ -201,7 +210,10 @@ spoAndCCVotingSpec = do
 
 committeeExpiryResignationDiscountSpec ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 committeeExpiryResignationDiscountSpec =
   -- Committee-update proposals are disallowed during bootstrap, so we can only run these tests post-bootstrap
@@ -273,7 +285,10 @@ committeeExpiryResignationDiscountSpec =
 
 paramChangeAffectsProposalsSpec ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 paramChangeAffectsProposalsSpec =
   -- These tests rely on submitting committee-update proposals and on drep votes, which are disallowed during bootstrap,
@@ -451,7 +466,10 @@ paramChangeAffectsProposalsSpec =
 
 committeeMinSizeAffectsInFlightProposalsSpec ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 committeeMinSizeAffectsInFlightProposalsSpec =
   -- Treasury withdrawals are disallowed during bootstrap, so we can only run these tests post-bootstrap
@@ -518,7 +536,10 @@ committeeMinSizeAffectsInFlightProposalsSpec =
 
 spoVotesForHardForkInitiation ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 spoVotesForHardForkInitiation =
   describe "Counting of SPO votes" $ do
@@ -556,7 +577,10 @@ spoVotesForHardForkInitiation =
 
 votingSpec ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 votingSpec =
   describe "Voting" $ do
@@ -1455,7 +1479,10 @@ votingSpec =
 
 delayingActionsSpec ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 delayingActionsSpec =
   -- All tests below are relying on submitting constitution of committe-update proposals, which are disallowed during bootstrap,
@@ -1672,7 +1699,10 @@ delayingActionsSpec =
 
 committeeMaxTermLengthSpec ::
   forall era.
-  (HasCallStack, ConwayEraImp era) =>
+  ( HasCallStack
+  , ConwayEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 committeeMaxTermLengthSpec =
   -- Committee-update proposals are disallowed during bootstrap, so we can only run these tests post-bootstrap