+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/ImpTest.hs
@@ -313,7 +313,6 @@ instance AlonzoEraImp ConwayEra where
 class
   ( AlonzoEraImp era
   , ConwayEraTest era
-  , ConwayEraTxCert era
   , STS (EraRule "ENACT" era)
   , BaseM (EraRule "ENACT" era) ~ ShelleyBase
   , State (EraRule "ENACT" era) ~ EnactState era
@@ -375,6 +374,7 @@ unRegisterDRep drep = do
 genUnRegTxCert ::
   forall era.
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxCert era
   ) =>
   Credential 'Staking ->
@@ -392,6 +392,7 @@ genUnRegTxCert stakingCredential = do
 genRegTxCert ::
   forall era.
   ( ShelleyEraImp era
+  , ShelleyEraTxCert era
   , ConwayEraTxCert era
   ) =>
   Credential 'Staking ->
@@ -489,12 +490,12 @@ getDRepState dRepCred = do
 -- in Conway. The Shelley version of this function would have to separately
 -- register the staking credential and then delegate it.
 setupPoolWithStake ::
-  (ShelleyEraImp era, ConwayEraTxCert era) =>
+  ConwayEraImp era =>
   Coin ->
   ImpTestM era (KeyHash 'StakePool, Credential 'Payment, Credential 'Staking)
 setupPoolWithStake delegCoin = do
   khPool <- freshKeyHash
-  registerPool khPool
+  registerPoolWithDeposit khPool
   credDelegatorPayment <- KeyHashObj <$> freshKeyHash
   credDelegatorStaking <- KeyHashObj <$> freshKeyHash
   sendCoinTo_ (mkAddr credDelegatorPayment credDelegatorStaking) delegCoin
@@ -511,11 +512,11 @@ setupPoolWithStake delegCoin = do
   pure (khPool, credDelegatorPayment, credDelegatorStaking)
 
 setupPoolWithoutStake ::
-  (ShelleyEraImp era, ConwayEraTxCert era) =>
+  ConwayEraImp era =>
   ImpTestM era (KeyHash 'StakePool, Credential 'Staking)
 setupPoolWithoutStake = do
   khPool <- freshKeyHash
-  registerPool khPool
+  registerPoolWithDeposit khPool
   credDelegatorStaking <- KeyHashObj <$> freshKeyHash
   deposit <- getsNES $ nesEsL . curPParamsEpochStateL . ppKeyDepositL
   submitTxAnn_ "Delegate to stake pool" $
@@ -685,9 +686,7 @@ submitFailingProposal proposal expectedFailure =
 -- | Submits a transaction that proposes the given governance action. For proposing
 -- multiple actions in the same transaciton use `trySubmitGovActions` instead.
 trySubmitGovAction ::
-  ( ShelleyEraImp era
-  , ConwayEraTxBody era
-  ) =>
+  ConwayEraImp era =>
   GovAction era ->
   ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) GovActionId)
 trySubmitGovAction ga = do
@@ -717,7 +716,7 @@ submitAndExpireProposalToMakeReward stakingC = do
 
 -- | Submits a transaction that proposes the given governance action
 trySubmitGovActions ::
-  (ShelleyEraImp era, ConwayEraTxBody era) =>
+  ConwayEraImp era =>
   NE.NonEmpty (GovAction era) ->
   ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era)), Tx era) (Tx era))
 trySubmitGovActions gas = do
@@ -741,17 +740,16 @@ mkProposalWithRewardAccount ga rewardAccount = do
       }
 
 mkProposal ::
-  (ShelleyEraImp era, ConwayEraTxBody era) =>
+  ConwayEraImp era =>
   GovAction era ->
   ImpTestM era (ProposalProcedure era)
 mkProposal ga = do
-  rewardAccount <- registerRewardAccount
+  rewardAccount <- registerRewardAccountWithDeposit
   mkProposalWithRewardAccount ga rewardAccount
 
 submitGovAction ::
   forall era.
-  ( ShelleyEraImp era
-  , ConwayEraTxBody era
+  ( ConwayEraImp era
   , HasCallStack
   ) =>
   GovAction era ->
@@ -762,8 +760,7 @@ submitGovAction ga = do
 
 submitGovAction_ ::
   forall era.
-  ( ShelleyEraImp era
-  , ConwayEraTxBody era
+  ( ConwayEraImp era
   , HasCallStack
   ) =>
   GovAction era ->
@@ -772,8 +769,7 @@ submitGovAction_ = void . submitGovAction
 
 submitGovActions ::
   forall era.
-  ( ShelleyEraImp era
-  , ConwayEraTxBody era
+  ( ConwayEraImp era
   , HasCallStack
   ) =>
   NE.NonEmpty (GovAction era) ->
@@ -791,10 +787,7 @@ mkTreasuryWithdrawalsGovAction wdrls =
   TreasuryWithdrawals (Map.fromList wdrls) <$> getGovPolicy
 
 submitTreasuryWithdrawals ::
-  ( ShelleyEraImp era
-  , ConwayEraTxBody era
-  , ConwayEraGov era
-  ) =>
+  ConwayEraImp era =>
   [(RewardAccount, Coin)] ->
   ImpTestM era GovActionId
 submitTreasuryWithdrawals wdrls =
@@ -844,8 +837,7 @@ getGovPolicy =
 
 submitFailingGovAction ::
   forall era.
-  ( ShelleyEraImp era
-  , ConwayEraTxBody era
+  ( ConwayEraImp era
   , HasCallStack
   ) =>
   GovAction era ->