+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -390,8 +390,7 @@ impEventsL :: Lens' (ImpTestState era) [SomeSTSEvent era]
 impEventsL = lens impEvents (\x y -> x {impEvents = y})
 
 class
-  ( ShelleyEraTxCert era
-  , ShelleyEraTest era
+  ( ShelleyEraTest era
   , -- For BBODY rule
     STS (EraRule "BBODY" era)
   , BaseM (EraRule "BBODY" era) ~ ShelleyBase
@@ -1466,6 +1465,7 @@ registerStakeCredential ::
   forall era.
   ( HasCallStack
   , ShelleyEraImp era
+  , ShelleyEraTxCert era
   ) =>
   Credential 'Staking ->
   ImpTestM era RewardAccount
@@ -1478,7 +1478,9 @@ registerStakeCredential cred = do
   pure $ RewardAccount networkId cred
 
 delegateStake ::
-  ShelleyEraImp era =>
+  ( ShelleyEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   Credential 'Staking ->
   KeyHash 'StakePool ->
   ImpTestM era ()
@@ -1493,6 +1495,7 @@ registerRewardAccount ::
   forall era.
   ( HasCallStack
   , ShelleyEraImp era
+  , ShelleyEraTxCert era
   ) =>
   ImpTestM era RewardAccount
 registerRewardAccount = do
@@ -1532,7 +1535,9 @@ freshPoolParams khPool rewardAccount = do
       }
 
 registerPool ::
-  ShelleyEraImp era =>
+  ( ShelleyEraImp era
+  , ShelleyEraTxCert era
+  ) =>
   KeyHash 'StakePool ->
   ImpTestM era ()
 registerPool khPool = registerRewardAccount >>= registerPoolWithRewardAccount khPool